codegen: clippy fixes + cargo fmt
* `(disp as u32) % 4 == 0` -> `(disp as u32).is_multiple_of(4)` in
  the aarch64 LoadLocal F32 path (clippy::manual_is_multiple_of
  promoted to error under `-D warnings` in CI).
* Redundant `(cond as u32)` cast in branch-fuse analyser dropped
  (cond is already u32).
* `cargo fmt` over the SSA codegen + builder modules to match the
  project's rustfmt config; touched seven files.

Unblocks CI on the cfg_ssa branch.

+++ src/c5/codegen/ssa_alloc.rs
@@ -36,10 +36,10 @@
 use alloc::vec;
 use alloc::vec::Vec;
 
-use super::Target;
 use super::super::ir::{
     BinOp, FpCastKind, FunctionSsa, Inst, LoadKind, NO_VALUE, StoreKind, ValueId,
 };
+use super::Target;
 
 /// Where the allocator placed an SSA value.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -379,7 +379,7 @@ pub(super) fn allocate(func: &FunctionSsa, target: Target) -> Allocation {
         if cond == NO_VALUE {
             continue;
         }
-        if (cond as u32) + 1 != block.inst_range.end {
+        if cond + 1 != block.inst_range.end {
             continue;
         }
         if use_counts.get(cond as usize).copied().unwrap_or(0) != 1 {
